Classify direct, ambiguous, or no effect between exposure and network nodes.

Using the output of the nc_exposure_estimates() or nc_outcome_estimates(), classify whether the exposure variable has a direct, ambiguous, or no link with the index node metabolic variable, conditional on neighbouring metablic variables and on potential confounders.

nc_classify_effects(.tbl)

nc_filter_estimates(.tbl)

Arguments

.tbl
Multi-model estimates generated from nc_exposure_estimates() or nc_outcome_estimates() that contain the model summaries.

Value

Outputs a tibble with model estimates between the exposure and the individual index network nodes, along with the classification of direct effect or not.

Functions

nc_filter_estimates: Filter out estimates so only exposure estimates are kept. Adds a column with the neighbouring metabolite variables listed.

See also

Examples

library(dplyr)
metabolite_network <- simulated_data %>%
  select(matches("metabolite")) %>%
  nc_create_network()
multimodel_exposure <- simulated_data %>%
  mutate(Random = rnorm(nrow(.))) %>%
  nc_exposure_estimates(
    .graph = metabolite_network,
    .exposure = "exposure",
    .adjustment_vars = c("age", "Random"),
    .model_function = lm
   )

nc_filter_estimates(multimodel_exposure)

#> # A tibble: 114 x 10
#>    exposure index_node estimate std_error statistic   p_value conf_low conf_high
#>    <chr>    <chr>         <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1 exposure metabolit…   -0.725    0.0157     -46.3 1.61e-318   -0.756    -0.694
#>  2 exposure metabolit…   -0.711    0.0253     -28.1 7.89e-147   -0.760    -0.661
#>  3 exposure metabolit…   -0.767    0.0140     -54.8 0.          -0.794    -0.739
#>  4 exposure metabolit…   -0.756    0.0242     -31.2 9.63e-175   -0.803    -0.708
#>  5 exposure metabolit…   -0.725    0.0258     -28.1 1.23e-146   -0.776    -0.674
#>  6 exposure metabolit…   -0.764    0.0155     -49.2 0.          -0.794    -0.734
#>  7 exposure metabolit…   -0.764    0.0247     -31.0 2.40e-172   -0.813    -0.716
#>  8 exposure metabolit…   -0.799    0.0139     -57.6 0.          -0.826    -0.771
#>  9 exposure metabolit…    0.284    0.0256      11.1 1.02e- 27    0.234     0.334
#> 10 exposure metabolit…    0.227    0.0156      14.6 1.04e- 45    0.197     0.258
#> # … with 104 more rows, and 2 more variables: adjusted_vars <chr>,
#> #   neighbour_vars <chr>

nc_classify_effects(multimodel_exposure)

#> # A tibble: 12 x 9
#> # Groups:   exposure, index_node [12]
#>    exposure index_node estimate std_error p_value conf_low conf_high adj_p_value
#>    <chr>    <chr>         <dbl>     <dbl>   <dbl>    <dbl>     <dbl>       <dbl>
#>  1 exposure metabolit… -0.799      0.0139 0        -0.826    -0.771      0      
#>  2 exposure metabolit…  0.494      0.0193 0         0.456     0.532      0      
#>  3 exposure metabolit…  0.290      0.0217 0         0.247     0.332      0      
#>  4 exposure metabolit… -0.462      0.0199 0        -0.502    -0.423      0      
#>  5 exposure metabolit…  0.247      0.0219 0         0.204     0.290      0      
#>  6 exposure metabolit…  0.0686     0.0224 0.00225   0.0246    0.113      0.00292
#>  7 exposure metabolit…  0.0214     0.0228 0.349    -0.0234    0.0662     0.388  
#>  8 exposure metabolit… -0.00708    0.0225 0.753    -0.0512    0.0370     0.760  
#>  9 exposure metabolit… -0.0211     0.0229 0.357    -0.0661    0.0239     0.392  
#> 10 exposure metabolit…  0.0643     0.0218 0.00325   0.0215    0.107      0.00412
#> 11 exposure metabolit…  0.835      0.0127 0         0.810     0.860      0      
#> 12 exposure metabolit…  0.375      0.0207 0         0.335     0.416      0      
#> # … with 1 more variable: direct_effect <chr>

library(survival)
multimodel_outcome <- simulated_data %>%
  mutate(Random = rnorm(nrow(.))) %>%
  nc_outcome_estimates(
    .graph = metabolite_network,
    .outcome = "survival::Surv(survival_time, case_status)",
    .adjustment_vars = c("age", "Random"),
    .model_function = survival::coxph,
    .exponentiate = TRUE
   )

nc_filter_estimates(multimodel_outcome)

#> # A tibble: 114 x 10
#>    outcome index_node estimate std_error statistic  p_value conf_low conf_high
#>    <chr>   <chr>         <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#>  1 surviv… metabolit…    0.847    0.103     -1.61  1.07e- 1    0.692     1.04 
#>  2 surviv… metabolit…    0.811    0.100     -2.09  3.68e- 2    0.666     0.987
#>  3 surviv… metabolit…    0.775    0.0978    -2.61  8.99e- 3    0.640     0.938
#>  4 surviv… metabolit…    0.753    0.0960    -2.96  3.10e- 3    0.624     0.909
#>  5 surviv… metabolit…    0.631    0.0847    -5.45  5.14e- 8    0.534     0.744
#>  6 surviv… metabolit…    0.591    0.0805    -6.53  6.44e-11    0.505     0.692
#>  7 surviv… metabolit…    0.556    0.0763    -7.71  1.27e-14    0.478     0.645
#>  8 surviv… metabolit…    0.530    0.0729    -8.73  2.65e-18    0.459     0.611
#>  9 surviv… metabolit…    1.13     0.102      1.20  2.29e- 1    0.926     1.38 
#> 10 surviv… metabolit…    1.10     0.109      0.887 3.75e- 1    0.890     1.36 
#> # … with 104 more rows, and 2 more variables: adjusted_vars <chr>,
#> #   neighbour_vars <chr>

nc_classify_effects(multimodel_outcome)

#> # A tibble: 12 x 9
#> # Groups:   outcome, index_node [12]
#>    outcome index_node estimate std_error p_value conf_low conf_high adj_p_value
#>    <chr>   <chr>         <dbl>     <dbl>   <dbl>    <dbl>     <dbl>       <dbl>
#>  1 surviv… metabolit…    0.530    0.0729 0.         0.459     0.611  0         
#>  2 surviv… metabolit…    1.48     0.0716 0.         1.28      1.70   0         
#>  3 surviv… metabolit…    1.27     0.0705 6.07e-4    1.11      1.46   0.00126   
#>  4 surviv… metabolit…    0.616    0.0726 0.         0.534     0.710  0         
#>  5 surviv… metabolit…    1.39     0.0715 4.00e-6    1.21      1.60   0.0000106 
#>  6 surviv… metabolit…    1.74     0.0716 0.         1.51      2.00   0         
#>  7 surviv… metabolit…    1.69     0.0700 0.         1.47      1.94   0         
#>  8 surviv… metabolit…    1.40     0.0712 2.00e-6    1.22      1.61   0.00000543
#>  9 surviv… metabolit…    0.863    0.0697 3.38e-2    0.753     0.989  0.0495    
#> 10 surviv… metabolit…    1.17     0.0728 3.10e-2    1.01      1.35   0.0460    
#> 11 surviv… metabolit…    2.03     0.0705 0.         1.77      2.33   0         
#> 12 surviv… metabolit…    2.32     0.0744 0.         2.01      2.69   0         
#> # … with 1 more variable: direct_effect <chr>