Create an estimate of the metabolic network as an undirected graph.

The main NetCoupler network creator. Uses the input data to estimate the underlying undirected graph. The default uses the PC algorithm, implemented within NetCoupler with pc_estimate_undirected_graph() Defaults to using the PC algorithm to calculate possible edges. Any missing values in the input data are removed by this function, since some computations can't handle missingness.

Usage

nc_estimate_network(data, cols = everything(), alpha = 0.01)

Arguments

data

Data that would form the underlying network.

cols

<tidy-select> Variables to include by using dplyr::select() style selection.

alpha

The alpha level to use to test whether an edge exists or not. Default is 0.01.

Value

Outputs a tidygraph::tbl_graph() with the start and end nodes, as well as the edge weights.

See also