Estimating pathways to an estimated DAG from an exposure.

This algorithm estimates direct effects of a predefined exposure on each network-variable for all causal models that agree with the input-network: models are adjusted for all possible combinations of direct neighbors (==variables in the adjacency set) -> Output is a multiset of possible effects

nc_exposure_estimates(
  .data,
  .graph,
  .exposure,
  .adjustment_vars,
  .model_function,
  .exponentiate = FALSE,
  ...
)

Arguments

.data
A data.frame that contains the data with the metabolic variables and the outcome.

.graph
The estimated graph skeleton obtained from nc_create_network().

.exposure
Character. The exposure variable of interest.

.adjustment_vars
Character vector. The variables to adjust for in the model.

.model_function
A function object. The function to use for the modeling.

.exponentiate
Logical. Default is false. Whether to exponentiate the model results output, usually necessary for e.g. logistic regression models.

...
Options to pass to the model function.

Value

Outputs a tibble with all the models computed and their estimates.

Examples

library(dplyr)
metabolite_network <- simulated_data %>%
  select(matches("metabolite")) %>%
  nc_create_network()
simulated_data %>%
  nc_exposure_estimates(
    .graph = metabolite_network,
    .exposure = "exposure",
    .adjustment_vars = "age",
    .model_function = lm
   )

#> # A tibble: 551 x 11
#>    exposure index_node model_id term  estimate std_error statistic   p_value
#>    <chr>    <chr>      <chr>    <chr>    <dbl>     <dbl>     <dbl>     <dbl>
#>  1 exposure metabolit… e600a8c… (Int…  6.26e+0   0.136      46.0   9.77e-316
#>  2 exposure metabolit… e600a8c… meta…  8.73e-2   0.0151      5.77  9.29e-  9
#>  3 exposure metabolit… e600a8c… meta… -1.36e-1   0.0138     -9.89  1.50e- 22
#>  4 exposure metabolit… e600a8c… age    6.88e-4   0.00137     0.501 6.16e-  1
#>  5 exposure metabolit… e600a8c… expo… -7.25e-1   0.0157    -46.3   2.53e-318
#>  6 exposure metabolit… cb4b27f… (Int…  6.36e+0   0.190      33.5   6.40e-196
#>  7 exposure metabolit… cb4b27f… meta…  8.76e-2   0.0151      5.79  8.24e-  9
#>  8 exposure metabolit… cb4b27f… meta… -1.35e-1   0.0139     -9.73  7.10e- 22
#>  9 exposure metabolit… cb4b27f… meta… -1.80e-2   0.0240     -0.750 4.54e-  1
#> 10 exposure metabolit… cb4b27f… age    6.67e-4   0.00137     0.486 6.27e-  1
#> # … with 541 more rows, and 3 more variables: conf_low <dbl>, conf_high <dbl>,
#> #   adjusted_vars <chr>