Compute model estimates between an external (exposure or outcome) variable and a network.

TODO: Describe more here (fill out vignette first).

nc_exposure_estimates(
  .tbl,
  .graph,
  .exposure,
  .adjustment_vars = NA,
  .model_function,
  .model_arg_list = NULL,
  .exponentiate = FALSE
)

nc_outcome_estimates(
  .tbl,
  .graph,
  .outcome,
  .adjustment_vars = NA,
  .model_function,
  .model_arg_list = NULL,
  .exponentiate = FALSE
)

.compute_model_estimates(
  .tbl,
  .graph,
  .external_var,
  .adjustment_vars = NA,
  .model_function,
  .model_arg_list = NULL,
  .exponentiate = FALSE,
  .external_side = c("exposure", "outcome")
)

Arguments

.tbl
The data.frame or tibble that contains the variables of interest, including the variables passed to the network.

.graph
Output graph object from nc_create_network().

.exposure, .outcome
Character. The exposure or outcome variable of interest.

.adjustment_vars
Optional. Variables to adjust for in the models.

.model_function
A function for the model to use (e.g. stats::lm(), stats::glm(), survival::coxph()). Can be any model as long as the function has the arguments formula and data.

.model_arg_list
Optional. A list containing the named arguments that will be passed to the model function. A simple example would be list(family = binomial(link = "logit")) to specify that the glm model is a logistic model and not a linear one. See the examples for more on the usage.

.exponentiate
Logical. Whether to exponentiate the log estimates, as computed with e.g. logistic regression models.

.external_var
Argument for internal function. The variable that links to the network variables ("external" to the network).

.external_side
Argument for internal function. Character vector. Either "exposure" or "outcome", to indicate which side the external variable is on relative to the network.

Exposure indicating the implied directionality is from the external variable to the network variable.

Outcome indicating the implied directionality is from the network variable to the external variable.

Value

Outputs a tibble that contains the model estimates from either the exposure or outcome side of the network.

Functions

nc_exposure_estimates: Computes the model estimates for the exposure side.

nc_outcome_estimates: Computes the model estimates for the exposure side.

.compute_model_estimates: Internal function. Included to document algorithm.

Examples

library(dplyr)
metabolite_network <- simulated_data %>%
  select(matches("metabolite")) %>%
  nc_create_network()
simulated_data %>%
  nc_exposure_estimates(
    .graph = metabolite_network,
    .exposure = "exposure",
    .adjustment_vars = "age",
    .model_function = lm
   )

#> # A tibble: 551 x 16
#>    exposure index_node model_id term  estimate std_error statistic   p_value
#>    <chr>    <chr>      <chr>    <chr>    <dbl>     <dbl>     <dbl>     <dbl>
#>  1 exposure metabolit… 0916568… (Int…  5.52e+0   0.116      47.4   0.       
#>  2 exposure metabolit… 0916568… meta…  7.60e-2   0.0154      4.92  9.47e-  7
#>  3 exposure metabolit… 0916568… age    3.66e-4   0.00141     0.260 7.95e-  1
#>  4 exposure metabolit… 0916568… expo… -7.63e-1   0.0155    -49.2   0.       
#>  5 exposure metabolit… 30eeebe… (Int…  6.74e+0   0.108      62.6   0.       
#>  6 exposure metabolit… 30eeebe… meta… -1.30e-1   0.0138     -9.41  1.36e- 20
#>  7 exposure metabolit… 30eeebe… age    7.06e-4   0.00138     0.510 6.10e-  1
#>  8 exposure metabolit… 30eeebe… expo… -7.66e-1   0.0140    -54.8   0.       
#>  9 exposure metabolit… 2dcf6a0… (Int…  6.22e+0   0.164      37.9   5.76e-237
#> 10 exposure metabolit… 2dcf6a0… meta… -4.12e-2   0.0245     -1.68  9.23e-  2
#> # … with 541 more rows, and 8 more variables: conf_low <dbl>, conf_high <dbl>,
#> #   df <int>, logLik <dbl>, AIC <dbl>, BIC <dbl>, sample_size <int>,
#> #   adjusted_vars <chr>

simulated_data %>%
  nc_outcome_estimates(
    .graph = metabolite_network,
    .outcome = "survival::Surv(survival_time, case_status)",
    .adjustment_vars = "age",
    .model_function = survival::coxph,
    .exponentiate = TRUE
   )

#> # A tibble: 437 x 11
#>    outcome index_node model_id term  estimate std_error statistic  p_value
#>    <chr>   <chr>      <chr>    <chr>    <dbl>     <dbl>     <dbl>    <dbl>
#>  1 surviv… metabolit… 8460189… meta…    0.770   0.0796      -3.28 1.03e- 3
#>  2 surviv… metabolit… 8460189… age      1.01    0.00712      1.39 1.63e- 1
#>  3 surviv… metabolit… 8460189… meta…    0.591   0.0804      -6.54 6.02e-11
#>  4 surviv… metabolit… ade51bf… meta…    1.17    0.0769       2.08 3.79e- 2
#>  5 surviv… metabolit… ade51bf… age      1.01    0.00711      1.49 1.36e- 1
#>  6 surviv… metabolit… ade51bf… meta…    0.556   0.0762      -7.71 1.26e-14
#>  7 surviv… metabolit… fd65955… meta…    1.69    0.0939       5.57 2.56e- 8
#>  8 surviv… metabolit… fd65955… age      1.01    0.00711      1.56 1.18e- 1
#>  9 surviv… metabolit… fd65955… meta…    0.751   0.0958      -2.98 2.84e- 3
#> 10 surviv… metabolit… a6293b2… meta…    0.755   0.0804      -3.50 4.65e- 4
#> # … with 427 more rows, and 3 more variables: conf_low <dbl>, conf_high <dbl>,
#> #   adjusted_vars <chr>