Estimating pathways to an outcome from an estimated DAG.

This algorithm estimates direct effect of a predefined exposure (network-variable) on time-to-event for all causal models that agree with the input-network: Cox prop. hazards regression models are used to estimate the efect of all network-variables on survival time adjusted for all possible combinations of direct neighbors (adjacency set) -> Output is a multiset of possible causal effects.

nc_outcome_estimates(
  .data,
  .graph,
  .outcome,
  .adjustment_vars,
  .model_function,
  .exponentiate = FALSE,
  ...
)

Arguments

.data
A data.frame that contains the data with the metabolic variables and the outcome.

.graph
The estimated graph skeleton obtained from nc_create_network().

.outcome
Character. The outcome variable of interest.

.adjustment_vars
Character vector. The variables to adjust for in the model.

.model_function
A function object. The function to use for the modeling.

.exponentiate
Logical. Default is false. Whether to exponentiate the model results output, usually necessary for e.g. logistic regression models.

...
Options to pass to the model function.

Value

Outputs a tibble with all the models computed and their estimates.

Examples

library(dplyr)
library(survival)
metabolite_network <- simulated_data %>%
  select(matches("metabolite")) %>%
  nc_create_network()
simulated_data %>%
  nc_outcome_estimates(
    .graph = metabolite_network,
    .outcome = "survival::Surv(survival_time, case_status)",
    .adjustment_vars = "age",
    .model_function = survival::coxph,
    .exponentiate = TRUE
   )

#> # A tibble: 437 x 11
#>    outcome index_node model_id term  estimate std_error statistic p_value
#>    <chr>   <chr>      <chr>    <chr>    <dbl>     <dbl>     <dbl>   <dbl>
#>  1 surviv… metabolit… 099f0d7… meta…    0.789   0.0813      -2.91 3.58e-3
#>  2 surviv… metabolit… 099f0d7… meta…    1.16    0.0779       1.92 5.43e-2
#>  3 surviv… metabolit… 099f0d7… meta…    1.61    0.0953       5.02 5.26e-7
#>  4 surviv… metabolit… 099f0d7… age      1.01    0.00711      1.45 1.47e-1
#>  5 surviv… metabolit… 099f0d7… meta…    0.844   0.103       -1.65 9.97e-2
#>  6 surviv… metabolit… fa2fb13… meta…    0.805   0.0803      -2.69 7.05e-3
#>  7 surviv… metabolit… fa2fb13… meta…    1.64    0.0951       5.22 1.75e-7
#>  8 surviv… metabolit… fa2fb13… age      1.01    0.00711      1.47 1.42e-1
#>  9 surviv… metabolit… fa2fb13… meta…    0.808   0.100       -2.13 3.35e-2
#> 10 surviv… metabolit… 4c9da01… meta…    1.13    0.0772       1.57 1.17e-1
#> # … with 427 more rows, and 3 more variables: conf_low <dbl>, conf_high <dbl>,
#> #   adjusted_vars <chr>