Estimate Pearson's partial correlation coefficients.

This function is a wrapper around ppcor::pcor() that extracts correlation coefficient estimates, then adds the variable names to the column and row names.

nc_partial_corr_matrix(.data)

Arguments

.data
Input data of metabolic variables as matrix or data.frame.

Value

Outputs a matrix of partial correlation coefficients.

Examples

library(dplyr)
simulated_data %>%
  select(contains("metabolite")) %>%
  nc_partial_corr_matrix()

#>               metabolite_1 metabolite_10 metabolite_11 metabolite_12
#> metabolite_1   1.000000000   0.055485627  -0.012440234   0.256753103
#> metabolite_10  0.055485627   1.000000000   0.309330503  -0.636230203
#> metabolite_11 -0.012440234   0.309330503   1.000000000  -0.015561677
#> metabolite_12  0.256753103  -0.636230203  -0.015561677   1.000000000
#> metabolite_2  -0.231454137  -0.008008672  -0.006320635  -0.008046791
#> metabolite_3  -0.170261469   0.414081034   0.031834773   0.638129595
#> metabolite_4   0.129284074  -0.306175814  -0.014743194  -0.422991562
#> metabolite_5  -0.004715791   0.028539886  -0.052334110  -0.025295455
#> metabolite_6   0.020929152   0.227127327  -0.036897720   0.023387742
#> metabolite_7   0.228066916  -0.058859462  -0.012036327   0.014924324
#> metabolite_8  -0.497871008   0.225067943   0.012841128  -0.005226102
#> metabolite_9  -0.005079062  -0.210571751   0.007199316  -0.016836501
#>                metabolite_2 metabolite_3 metabolite_4  metabolite_5
#> metabolite_1  -0.2314541368 -0.170261469   0.12928407 -0.0047157909
#> metabolite_10 -0.0080086719  0.414081034  -0.30617581  0.0285398861
#> metabolite_11 -0.0063206352  0.031834773  -0.01474319 -0.0523341104
#> metabolite_12 -0.0080467908  0.638129595  -0.42299156 -0.0252954548
#> metabolite_2   1.0000000000  0.136414449   0.01247093 -0.0008961695
#> metabolite_3   0.1364144486  1.000000000   0.59582099  0.0347459616
#> metabolite_4   0.0124709295  0.595820986   1.00000000  0.1517525602
#> metabolite_5  -0.0008961695  0.034745962   0.15175256  1.0000000000
#> metabolite_6  -0.0074160741 -0.007441582   0.03280608 -0.4685247506
#> metabolite_7   0.2685238132 -0.007529180  -0.01989338 -0.0082598663
#> metabolite_8  -0.0261073031 -0.010075612   0.05739808 -0.2442623945
#> metabolite_9   0.0006178725  0.034098843  -0.05340990  0.4947203556
#>               metabolite_6 metabolite_7 metabolite_8  metabolite_9
#> metabolite_1   0.020929152  0.228066916 -0.497871008 -0.0050790622
#> metabolite_10  0.227127327 -0.058859462  0.225067943 -0.2105717505
#> metabolite_11 -0.036897720 -0.012036327  0.012841128  0.0071993164
#> metabolite_12  0.023387742  0.014924324 -0.005226102 -0.0168365009
#> metabolite_2  -0.007416074  0.268523813 -0.026107303  0.0006178725
#> metabolite_3  -0.007441582 -0.007529180 -0.010075612  0.0340988433
#> metabolite_4   0.032806083 -0.019893384  0.057398081 -0.0534098968
#> metabolite_5  -0.468524751 -0.008259866 -0.244262394  0.4947203556
#> metabolite_6   1.000000000  0.048016413 -0.117297841  0.0665501294
#> metabolite_7   0.048016413  1.000000000  0.344735648 -0.0182426709
#> metabolite_8  -0.117297841  0.344735648  1.000000000  0.4978454519
#> metabolite_9   0.066550129 -0.018242671  0.497845452  1.0000000000