Plot of the network of the metabolic variables.

nc_plot_network(
  .tbl,
  .graph,
  .edge_label_threshold = 0.2,
  .fn_node_rename = NULL
)

Arguments

.tbl
The data that was also used to generate the nc_estimate_network().

.graph
The network graph object of the metabolic variable network.

.edge_label_threshold
Threshold set for edge weight value above which the edge label will be kept. This argument helps to reduce clutter in the graph.

.fn_node_rename
Function to pass to rename the metabolic variables. Preferably use functions that search and replace patterns, like gsub() or stringr::str_replace().

Value

Outputs a ggplot2 object of the metabolic network.

See also