Classify direct, ambiguous, or no effect between exposure and network nodes.

Using the output of the nc_exposure_estimates() or nc_outcome_estimates(), classify whether the exposure variable has a direct, ambiguous, or no link with the index node metabolic variable, conditional on neighbouring metablic variables and on potential confounders.

nc_classify_effects(.tbl)

nc_filter_estimates(.tbl)

Arguments

.tbl
Multi-model estimates generated from nc_exposure_estimates() or nc_outcome_estimates() that contain the model summaries.

Value

Outputs a tibble with model estimates between the exposure and the individual index network nodes, along with the classification of direct effect or not.

Functions

nc_filter_estimates: Filter out estimates so only exposure estimates are kept. Adds a column with the neighbouring metabolite variables listed.

See also

Examples

library(dplyr)
metabolite_network <- simulated_data %>%
  select(matches("metabolite")) %>%
  nc_create_network()
multimodel_exposure <- simulated_data %>%
  mutate(Random = rnorm(nrow(.)),
         Sex = sample(rep(c("F", "M"), times = nrow(.) / 2))) %>%
  nc_exposure_estimates(
    .graph = metabolite_network,
    .exposure = "exposure",
    .adjustment_vars = c("age", "Random", "Sex"),
    .model_function = lm
   )

nc_filter_estimates(multimodel_exposure)

#> # A tibble: 114 x 15
#>    exposure index_node estimate std_error statistic   p_value conf_low conf_high
#>    <chr>    <chr>         <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1 exposure metabolit…   -0.764    0.0155     -49.2 0.          -0.794    -0.734
#>  2 exposure metabolit…   -0.767    0.0140     -54.8 0.          -0.794    -0.739
#>  3 exposure metabolit…   -0.765    0.0247     -31.0 2.66e-172   -0.813    -0.716
#>  4 exposure metabolit…   -0.725    0.0157     -46.3 2.20e-318   -0.756    -0.694
#>  5 exposure metabolit…   -0.725    0.0258     -28.1 1.31e-146   -0.776    -0.675
#>  6 exposure metabolit…   -0.756    0.0242     -31.2 1.10e-174   -0.803    -0.708
#>  7 exposure metabolit…   -0.711    0.0253     -28.1 8.57e-147   -0.760    -0.661
#>  8 exposure metabolit…   -0.799    0.0139     -57.6 0.          -0.826    -0.771
#>  9 exposure metabolit…    0.379    0.0180      21.0 1.28e- 88    0.343     0.414
#> 10 exposure metabolit…    0.231    0.0176      13.1 9.47e- 38    0.197     0.266
#> # … with 104 more rows, and 7 more variables: df <int>, logLik <dbl>,
#> #   AIC <dbl>, BIC <dbl>, sample_size <int>, adjusted_vars <chr>,
#> #   neighbour_vars <chr>

nc_classify_effects(multimodel_exposure)

#> # A tibble: 12 x 14
#>    exposure index_node estimate std_error p_value conf_low conf_high    df
#>    <chr>    <chr>         <dbl>     <dbl>   <dbl>    <dbl>     <dbl> <int>
#>  1 exposure metabolit… -0.799      0.0139 0        -0.826    -0.771      5
#>  2 exposure metabolit…  0.494      0.0193 0         0.456     0.532      5
#>  3 exposure metabolit…  0.290      0.0218 0         0.247     0.332      5
#>  4 exposure metabolit… -0.462      0.0199 0        -0.502    -0.423      5
#>  5 exposure metabolit…  0.247      0.0219 0         0.204     0.290      5
#>  6 exposure metabolit…  0.0686     0.0224 0.00227   0.0246    0.113      5
#>  7 exposure metabolit…  0.0214     0.0229 0.349    -0.0234    0.0662     5
#>  8 exposure metabolit… -0.00702    0.0225 0.755    -0.0511    0.0371     5
#>  9 exposure metabolit… -0.0211     0.0229 0.357    -0.0661    0.0239     5
#> 10 exposure metabolit…  0.0642     0.0218 0.00328   0.0214    0.107      5
#> 11 exposure metabolit…  0.835      0.0127 0         0.810     0.860      5
#> 12 exposure metabolit…  0.375      0.0207 0         0.335     0.416      5
#> # … with 6 more variables: logLik <dbl>, AIC <dbl>, BIC <dbl>,
#> #   sample_size <int>, adj_p_value <dbl>, direct_effect <chr>

library(survival)
multimodel_outcome <- simulated_data %>%
  mutate(Random = rnorm(nrow(.))) %>%
  nc_outcome_estimates(
    .graph = metabolite_network,
    .outcome = "survival::Surv(survival_time, case_status)",
    .adjustment_vars = c("age", "Random"),
    .model_function = survival::coxph,
    .exponentiate = TRUE
   )

nc_filter_estimates(multimodel_outcome)

#> # A tibble: 114 x 10
#>    outcome index_node estimate std_error statistic  p_value conf_low conf_high
#>    <chr>   <chr>         <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#>  1 surviv… metabolit…    0.589    0.0808     -6.55 5.63e-11    0.503     0.690
#>  2 surviv… metabolit…    0.555    0.0765     -7.71 1.25e-14    0.477     0.644
#>  3 surviv… metabolit…    0.750    0.0963     -2.98 2.85e- 3    0.621     0.906
#>  4 surviv… metabolit…    0.628    0.0849     -5.48 4.33e- 8    0.532     0.742
#>  5 surviv… metabolit…    0.806    0.100      -2.15 3.16e- 2    0.662     0.981
#>  6 surviv… metabolit…    0.773    0.0981     -2.63 8.60e- 3    0.638     0.937
#>  7 surviv… metabolit…    0.842    0.103      -1.67 9.53e- 2    0.688     1.03 
#>  8 surviv… metabolit…    0.528    0.0732     -8.72 2.85e-18    0.458     0.610
#>  9 surviv… metabolit…    1.42     0.0828      4.26 2.06e- 5    1.21      1.67 
#> 10 surviv… metabolit…    1.12     0.0979      1.12 2.62e- 1    0.921     1.35 
#> # … with 104 more rows, and 2 more variables: adjusted_vars <chr>,
#> #   neighbour_vars <chr>

nc_classify_effects(multimodel_outcome)

#> # A tibble: 12 x 9
#>    outcome index_node estimate std_error p_value conf_low conf_high adj_p_value
#>    <chr>   <chr>         <dbl>     <dbl>   <dbl>    <dbl>     <dbl>       <dbl>
#>  1 surviv… metabolit…    0.528    0.0732 0.         0.458     0.610  0         
#>  2 surviv… metabolit…    1.47     0.0716 0.         1.28      1.69   0         
#>  3 surviv… metabolit…    1.27     0.0706 7.00e-4    1.11      1.46   0.00145   
#>  4 surviv… metabolit…    0.619    0.0725 0.         0.537     0.713  0         
#>  5 surviv… metabolit…    1.39     0.0713 3.00e-6    1.21      1.60   0.00000795
#>  6 surviv… metabolit…    1.74     0.0717 0.         1.52      2.01   0         
#>  7 surviv… metabolit…    1.70     0.0704 0.         1.49      1.96   0         
#>  8 surviv… metabolit…    1.41     0.0715 1.00e-6    1.23      1.63   0.00000278
#>  9 surviv… metabolit…    0.858    0.0698 2.87e-2    0.749     0.984  0.0425    
#> 10 surviv… metabolit…    1.17     0.0727 3.44e-2    1.01      1.34   0.0485    
#> 11 surviv… metabolit…    2.03     0.0708 0.         1.77      2.33   0         
#> 12 surviv… metabolit…    2.34     0.0750 0.         2.02      2.71   0         
#> # … with 1 more variable: direct_effect <chr>