Compute model estimates between an external (exposure or outcome) variable and a network.

TODO: Describe more here (fill out vignette first).

nc_estimate_exposure_links(): Computes the model estimates for the exposure side.

nc_estimate_outcome_links(): Computes the model estimates for the exposure side.

nc_estimate_exposure_links(
  data,
  edge_tbl,
  exposure,
  adjustment_vars = NA,
  model_function,
  model_arg_list = NULL,
  exponentiate = FALSE,
  implementation = c("updated", "original")
)

nc_estimate_outcome_links(
  data,
  edge_tbl,
  outcome,
  adjustment_vars = NA,
  model_function,
  model_arg_list = NULL,
  exponentiate = FALSE,
  implementation = c("updated", "original")
)

Arguments

data
The data.frame or tibble that contains the variables of interest, including the variables used to make the network.

edge_tbl
Output graph object from nc_estimate_network(), converted to an edge table using as_edge_tbl().

exposure, outcome
Character. The exposure or outcome variable of interest.

adjustment_vars
Optional. Variables to adjust for in the models.

model_function
A function for the model to use (e.g. stats::lm(), stats::glm(), survival::coxph()). Can be any model as long as the function has the arguments formula and data.

model_arg_list
Optional. A list containing the named arguments that will be passed to the model function. A simple example would be list(family = binomial(link = "logit")) to specify that the glm model is a logistic model and not a linear one. See the examples for more on the usage.

exponentiate
Logical. Whether to exponentiate the log estimates, as computed with e.g. logistic regression models.

external_var
Argument for internal function. The variable that links to the network variables ("external" to the network).

external_side
Argument for internal function. Character vector. Either "exposure" or "outcome", to indicate which side the external variable is on relative to the network.

Exposure indicating the implied directionality is from the external variable to the network variable.

Outcome indicating the implied directionality is from the network variable to the external variable.

Value

Outputs a tibble that contains the model estimates from either the exposure or outcome side of the network.

Examples

if (FALSE) {
library(dplyr)
standardized_data <- simulated_data %>%
    nc_standardize(starts_with("metabolite"))

metabolite_network <- simulated_data %>%
    nc_standardize(starts_with("metabolite"),
                   regressed_on = "age") %>%
    nc_estimate_network(starts_with("metabolite"))
edge_table <- as_edge_tbl(metabolite_network)

standardized_data %>%
  nc_estimate_exposure_links(
    edge_tbl = edge_table,
    exposure = "exposure",
    model_function = lm
   )

# Using a logistic regression
standardized_data %>%
  nc_estimate_outcome_links(
    edge_tbl = edge_table,
    outcome = "case_status",
    model_function = glm,
    adjustment_vars = "age",
    model_arg_list = list(family = binomial(link = "logit")),
    exponentiate = TRUE
  )

# Adding random confounders to adjust for
standardized_data %>%
  mutate(Random = rnorm(nrow(.)),
         Sex = sample(rep(c("F", "M"), times = nrow(.) / 2))) %>%
  nc_estimate_exposure_links(
    edge_tbl = edge_table,
    exposure = "exposure",
    adjustment_vars = c("age", "Random", "Sex"),
    model_function = lm
   )
}