Compute model estimates between an external (exposure or outcome) variable and a network.

This is the main function that identifies potential links between external factors and the network. There are two functions to estimate and classify links:

nc_estimate_exposure_links(): Computes the model estimates for the exposure side.

nc_estimate_outcome_links(): Computes the model estimates for the exposure side.

nc_estimate_exposure_links(
  data,
  edge_tbl,
  exposure,
  adjustment_vars = NA,
  model_function,
  model_arg_list = NULL,
  exponentiate = FALSE,
  classify_option_list = classify_options()
)

nc_estimate_outcome_links(
  data,
  edge_tbl,
  outcome,
  adjustment_vars = NA,
  model_function,
  model_arg_list = NULL,
  exponentiate = FALSE,
  classify_option_list = classify_options()
)

Arguments

data

The data.frame or tibble that contains the variables of interest, including the variables used to make the network.

edge_tbl

Output graph object from nc_estimate_network(), converted to an edge table using as_edge_tbl().

exposure, outcome

Character. The exposure or outcome variable of interest.

adjustment_vars

Optional. Variables to adjust for in the models.

model_function

A function for the model to use (e.g. stats::lm(), stats::glm(), survival::coxph()). Can be any model as long as the function has the arguments formula and data. Type in the model function as a bare object (without (), for instance as lm).

model_arg_list

Optional. A list containing the named arguments that will be passed to the model function. A simple example would be list(family = binomial(link = "logit")) to specify that the glm model is a logistic model and not a linear one. See the examples for more on the usage.

exponentiate

Logical. Whether to exponentiate the log estimates, as computed with e.g. logistic regression models.

classify_option_list

A list with classification options for direct, ambigious, or no effects. Used with the classify_options() function with the arguments:

single_metabolite_threshold: Default of 0.05. P-values from models with only the index metabolite (no neighbour adjustment) are classified as effects if below this threshold. For larger sample sizes and networks, we recommend lowering the threshold to reduce risk of false positives.

network_threshold: Default of 0.1. P-values from any models that have direct neighbour adjustments are classified as effects if below this threshold. This is assumed as a one-sided p-value threshold. Like the threshold above, a lower value should be used for larger sample sizes and networks.

direct_effect_adjustment: Default is NA. After running the algorithm once, sometimes it's useful to adjust for the direct effects identified to confirm whether other links exist.

Value

Outputs a tibble that contains the model estimates from either the exposure or outcome side of the network as well as the effect classification. Each row represents the "no neighbour node adjusted" model and has the results for the outcome/exposure to index node pathway. Columns for the outcome are:

outcome or exposure: The name of the variable used as the external variable.

index_node: The name of the metabolite used as the index node from the network. In combination with the outcome/exposure variable, they represent the individual model used for the classification.

estimate: The estimate from the outcome/exposure and index node model.

std_error: The standard error from the outcome/exposure and index node model.

fdr_p_value: The False Discovery Rate-adjusted p-value from the outcome/exposure and index node model.

effect: The NetCoupler classified effect between the index node and the outcome/exposure. Effects are classified as "direct" (there is a probable link based on the given thresholds), "ambigious" (there is a potential link but not all thresholds were passed), and "none" (no potential link seen).

The tibble output also has an attribute that contains all the models generated before classification. Access it with attr(output, "all_models_df").

See also

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
standardized_data <- simulated_data %>%
    nc_standardize(starts_with("metabolite"))

metabolite_network <- simulated_data %>%
    nc_standardize(starts_with("metabolite"),
                   regressed_on = "age") %>%
    nc_estimate_network(starts_with("metabolite"))
edge_table <- as_edge_tbl(metabolite_network)

results <- standardized_data %>%
  nc_estimate_exposure_links(
    edge_tbl = edge_table,
    exposure = "exposure",
    model_function = lm
   )
results
#> # A tibble: 12 × 6
#>    exposure index_node    estimate std_error fdr_p_value effect   
#>  * <chr>    <chr>            <dbl>     <dbl>       <dbl> <chr>    
#>  1 exposure metabolite_1   0.173      0.0228      0      direct   
#>  2 exposure metabolite_10  0.318      0.0219      0      direct   
#>  3 exposure metabolite_11  0.0543     0.0232      0.0409 ambiguous
#>  4 exposure metabolite_12  0.0242     0.0231      0.380  none     
#>  5 exposure metabolite_2  -0.0430     0.0231      0.106  ambiguous
#>  6 exposure metabolite_3   0.0411     0.0231      0.123  ambiguous
#>  7 exposure metabolite_4   0.00344    0.0232      0.920  none     
#>  8 exposure metabolite_5   0.0479     0.0232      0.0717 ambiguous
#>  9 exposure metabolite_6  -0.0189     0.0230      0.506  none     
#> 10 exposure metabolite_7  -0.162      0.0229      0      direct   
#> 11 exposure metabolite_8  -0.355      0.0216      0      direct   
#> 12 exposure metabolite_9   0.0571     0.0230      0.0292 ambiguous

# Get results of all models used prior to classification
attr(results, "all_models_df")
#> # A tibble: 353 × 10
#>    exposure index_node   model_id    term  estimate std_error statistic  p_value
#>    <chr>    <chr>        <chr>       <chr>    <dbl>     <dbl>     <dbl>    <dbl>
#>  1 exposure metabolite_1 fe2b8fdf75… (Int…  0.00108    0.0219    0.0493 9.61e- 1
#>  2 exposure metabolite_1 fe2b8fdf75… expo…  0.181      0.0221    8.18   5.05e-16
#>  3 exposure metabolite_1 fe2b8fdf75… meta…  0.275      0.0219   12.5    1.35e-34
#>  4 exposure metabolite_1 d6e3403a88… (Int…  0.00330    0.0229    0.144  8.86e- 1
#>  5 exposure metabolite_1 d6e3403a88… expo…  0.166      0.0243    6.85   9.97e-12
#>  6 exposure metabolite_1 d6e3403a88… meta…  0.0195     0.0241    0.811  4.18e- 1
#>  7 exposure metabolite_1 7dad77d9a9… (Int…  0.00260    0.0222    0.117  9.07e- 1
#>  8 exposure metabolite_1 7dad77d9a9… expo…  0.171      0.0225    7.60   4.53e-14
#>  9 exposure metabolite_1 7dad77d9a9… meta…  0.187      0.0223    8.38   1.00e-16
#> 10 exposure metabolite_1 7351d4f8fc… (Int… -0.00278    0.0223   -0.125  9.01e- 1
#> # … with 343 more rows, and 2 more variables: network_neighbours <list>,
#> #   adjusted_vars <chr>

# Using a logistic regression
standardized_data %>%
  nc_estimate_outcome_links(
    edge_tbl = edge_table,
    outcome = "outcome_binary",
    model_function = glm,
    adjustment_vars = "age",
    model_arg_list = list(family = binomial(link = "logit")),
    exponentiate = TRUE
  )
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: algorithm did not converge
#> # A tibble: 12 × 6
#>    outcome        index_node    estimate std_error fdr_p_value effect
#>  * <chr>          <chr>            <dbl>     <dbl>       <dbl> <chr> 
#>  1 outcome_binary metabolite_1      1.00     8172.           1 none  
#>  2 outcome_binary metabolite_10     1.00     8187.           1 none  
#>  3 outcome_binary metabolite_11     1.00     8176.           1 none  
#>  4 outcome_binary metabolite_12     1.00     8172.           1 none  
#>  5 outcome_binary metabolite_2      1.00     8196.           1 none  
#>  6 outcome_binary metabolite_3      1.00     8128.           1 none  
#>  7 outcome_binary metabolite_4      1.00     8206.           1 none  
#>  8 outcome_binary metabolite_5      1.00     8191.           1 none  
#>  9 outcome_binary metabolite_6      1.00     8208.           1 none  
#> 10 outcome_binary metabolite_7      1.00     8248.           1 none  
#> 11 outcome_binary metabolite_8      1.00     8165.           1 none  
#> 12 outcome_binary metabolite_9      1.00     8215.           1 none  

# Adding random confounders to adjust for
standardized_data %>%
  mutate(Random = rnorm(nrow(.)),
         Sex = sample(rep(c("F", "M"), times = nrow(.) / 2))) %>%
  nc_estimate_exposure_links(
    edge_tbl = edge_table,
    exposure = "exposure",
    adjustment_vars = c("age", "Random", "Sex"),
    model_function = lm
   )
#> # A tibble: 12 × 6
#>    exposure index_node    estimate std_error fdr_p_value effect   
#>  * <chr>    <chr>            <dbl>     <dbl>       <dbl> <chr>    
#>  1 exposure metabolite_1   0.172      0.0230      0      direct   
#>  2 exposure metabolite_10  0.322      0.0222      0      direct   
#>  3 exposure metabolite_11  0.0542     0.0235      0.0443 ambiguous
#>  4 exposure metabolite_12  0.0290     0.0234      0.287  none     
#>  5 exposure metabolite_2  -0.0467     0.0234      0.0821 ambiguous
#>  6 exposure metabolite_3   0.0432     0.0234      0.109  ambiguous
#>  7 exposure metabolite_4   0.00490    0.0234      0.878  none     
#>  8 exposure metabolite_5   0.0437     0.0235      0.108  ambiguous
#>  9 exposure metabolite_6  -0.0183     0.0232      0.514  none     
#> 10 exposure metabolite_7  -0.159      0.0232      0      direct   
#> 11 exposure metabolite_8  -0.353      0.0218      0      direct   
#> 12 exposure metabolite_9   0.0566     0.0233      0.0333 ambiguous

# Adjusting for a previous direct effect
standardized_data %>%
  nc_estimate_exposure_links(
    edge_tbl = edge_table,
    exposure = "exposure",
    model_function = lm,
    adjustment_vars = "age",
    classify_option_list = classify_options(direct_effect_adjustment = c("metabolite_1"))
  )
#> # A tibble: 11 × 6
#>    exposure index_node    estimate std_error fdr_p_value effect   
#>  * <chr>    <chr>            <dbl>     <dbl>       <dbl> <chr>    
#>  1 exposure metabolite_10  0.318      0.0228    0        direct   
#>  2 exposure metabolite_11  0.0518     0.0240    0.0655   ambiguous
#>  3 exposure metabolite_12 -0.00287    0.0235    0.936    none     
#>  4 exposure metabolite_2  -0.0909     0.0230    0.000258 direct   
#>  5 exposure metabolite_3   0.0244     0.0237    0.424    ambiguous
#>  6 exposure metabolite_4  -0.00125    0.0240    0.958    none     
#>  7 exposure metabolite_5   0.0437     0.0241    0.129    ambiguous
#>  8 exposure metabolite_6  -0.0226     0.0239    0.463    none     
#>  9 exposure metabolite_7  -0.166      0.0236    0        direct   
#> 10 exposure metabolite_8  -0.359      0.0223    0        direct   
#> 11 exposure metabolite_9   0.0524     0.0237    0.0611   ambiguous