Standardize the metabolic variables.

Can standardize by either 1) log()-transforming and then applying scale() (mean-center and scaled by standard deviation), or 2) if .regressed_on variables are given, then log-transforming, running a linear regression to obtain the stats::residuals(), and finally scaled. Use .regressed_on to try to remove influence of potential confounding.

nc_standardize(.tbl, .vars, .regressed_on = NULL)

Arguments

.tbl
Data frame.

.vars
Metabolic variables that will make up the network.

.regressed_on
Optional. A character vector of variables to regress the metabolic variables on. Use if you want to standardize the metabolic variables on variables that are known to influence them, e.g. sex or age. Calculates the residuals from a linear regression model.

Value

Outputs a tibble object, with the original metabolic variables now standardized.

Examples

# Don't regress on any variable
simulated_data %>%
  nc_standardize(vars(matches("metabolite_")))

#> # A tibble: 2,000 x 16
#>    survival_time case_status exposure metabolite_1 metabolite_10 metabolite_11
#>            <dbl>       <dbl>    <dbl>        <dbl>         <dbl>         <dbl>
#>  1         0.635           0    1.12       -0.231         0.674         0.620 
#>  2         0.635           0    0.278      -0.452        -0.330         0.248 
#>  3         0.414           1   -0.266       0.294         0.695         0.453 
#>  4         0.635           0    0.115      -0.659        -0.0922        0.937 
#>  5         0.635           0    0.411      -2.00         -0.135        -0.411 
#>  6         0.635           0    0.624       0.0415       -0.115        -0.0466
#>  7         0.635           0    0.151      -0.0811       -0.120         0.694 
#>  8         0.635           0    1.79       -0.241         1.64          0.714 
#>  9         0.635           0   -0.379       1.19         -0.287        -1.37  
#> 10         0.635           0   -1.25        0.782        -1.90         -2.26  
#> # … with 1,990 more rows, and 10 more variables: metabolite_12 <dbl>,
#> #   metabolite_2 <dbl>, metabolite_3 <dbl>, metabolite_4 <dbl>,
#> #   metabolite_5 <dbl>, metabolite_6 <dbl>, metabolite_7 <dbl>,
#> #   metabolite_8 <dbl>, metabolite_9 <dbl>, age <dbl>

# Extract residuals by regressing on a variable
simulated_data %>%
  nc_standardize(vars(matches("metabolite_")), "age")

#> # A tibble: 2,000 x 16
#>    survival_time case_status exposure metabolite_1 metabolite_10 metabolite_11
#>            <dbl>       <dbl>    <dbl>        <dbl>         <dbl>         <dbl>
#>  1         0.635           0    1.12        1.17           1.62         1.58  
#>  2         0.635           0    0.278      -0.525         -0.479       -0.208 
#>  3         0.414           1   -0.266       0.265          0.452        0.343 
#>  4         0.635           0    0.115      -0.485         -0.231        0.249 
#>  5         0.635           0    0.411      -1.88          -1.05        -1.17  
#>  6         0.635           0    0.624       1.38           1.35         1.36  
#>  7         0.635           0    0.151      -0.258         -0.282        0.0986
#>  8         0.635           0    1.79        0.376          1.25         0.824 
#>  9         0.635           0   -0.379      -0.0656        -0.758       -1.25  
#> 10         0.635           0   -1.25        0.401         -0.826       -1.00  
#> # … with 1,990 more rows, and 10 more variables: metabolite_12 <dbl>,
#> #   metabolite_2 <dbl>, metabolite_3 <dbl>, metabolite_4 <dbl>,
#> #   metabolite_5 <dbl>, metabolite_6 <dbl>, metabolite_7 <dbl>,
#> #   metabolite_8 <dbl>, metabolite_9 <dbl>, age <dbl>